+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- CASMCMS-8624 - Adding default `kernel_file_name` based on arch type in the Job schema.
 
 ## [3.9.2] - 2023-05-19
 ### Changed